requirements.txt
grpcio==1.54.2
grpcio-tools==1.54.2
protobuf==4.23.2
pycrypto==2.6.1
pycryptodome==3.18.0
